feat: Modify chart symbols for SDF usage
Rendering of symbols is now based on SDF (signed distance field) which
only cares about the edges of symbol. Fill and stroke width in the
context of rendering, but now set to a some sensible values for easy
editing in Inkscape.

msdfgen used for SDF is not a full featured SVG parser. The symbols
are simplified as much as possible to leave only the essential
information.