feat: Remove clip margin for point items
This was originally used to render symbols that would cross tile
boundaries. Adjacent tiles would render their respective part of a
symbol onto their QImage which were quilted together in the final render.

With a scene graph based approach the tile boundaries are topological,
and not pixel based. Therefore a symbol should only exists once in the
final output.