feat: Define polygon holes in data model
With QPainter this wasn't that important as the winding order would
determine what areas would be filled in a polygon. But with tessellation
algorithms this is more important to know as holes may not cross the
main contour.

The capnp format has been updated to define polygon as a main line and
possibly several hole lines. A line in this terminology is the list of
points that defines a polygon or a hole.

Geometries in the S-57 objects are parsed so that the first geometry is
set as the main polygon and the rest are holes. This seems to work so
far, and hopefully the specification states it.