refactor: Major refactor scene graph nodes
The first implementation of this was a bit awkward with too much going
on in the QSGNode based classes.

With more QSGGeometry classes closer to the root node and using shared
QSGMaterial there is more potential for batching of draw calls.

The code is hopefully now a better of example of how to use the Qt scene
graph.

Summary of changes:

- Share QSGMaterial instances between tiles by letting RootNode create
  and destroy it.
- Managing of Tessellators now happens in Scene.
- QSGeometry based nodes are created once data is available to avoid
  creating a node with zero vertices first (problems with updating data
  later).
- Use a struct to define vertex layout.
- Tessellator::fetchData may outlive the Scene so SymbolImage and FontImage
  are now passed as shared_ptr.