feat: Reduce clipping margin
How to choose the ideal clipping margin is still work in progress.

The brief explanation is that for now this change is mainly to improve
the appearance of polygons with the upcoming scene graph rendering of
polygons.

The scene graph render code does not clip at tile level like maptile.cpp
does. The previously large clipping margin is also used in
ChartClipper::inflateAtEdges(). To avoid showing the weird artifacts
that inflateAtEdges introduce the clipping margin is reduced.

As an additional note ChartClipper::inflateAtEdges() might be an
unnecessary side step. At least with regards to land polygons. Their
polygon stroke mostly coincides line geometries such as "Coast Line"
from S-57. As soon as line clipping is added (Clipper only work for
polygons). This approach can be re-evaluated.