feat: Improve focus handling and remove Drawer functionality
The Drawer component did not work satisfactory on a tablet computer.
Drawer is enabled using swipe gesture which is a fragile functionality
across devices.

The focus handling is has been improved so that handlers are kept closer
to their relevant domains. There is no key handling property at Window
level so a new Item is inserted at root for that purpose.