feat: New debug UI for tiles
The QPainter based renderer has built-in debugging UI implemented in
the QML layout on top of the tile graphics. Because this part of the
code is being superseded by scene.cpp there is a need to re-implemen
the tile debugging UI.

This commit adds a sidebar on the left side that shows the relevant maps
for a tile and the possibility to toggle visibility of the tile. The
currently selected tile is highlighted with an translucent rectangle.