refactor: Replace QVector3D viewport with lat, lon and pixelsPerLon
Using QVector3D to pass those values between C++ and QML is confusing
as they do not represent an actual vector. This is a slight improvement
to the readability of the code. It is probably a good idea to soon
consider panning/zooming in mercator space instead of this awkward top-left +
pixelsPerLon regime.