feat: Calculate combined coverage to minimize rendered charts
How to determine which charts to render for a particular tile has been
difficult. The previous algorithm determined if tile bounding box was
completely enclosed by the chart bounding box.

This new approach uses Clipper2 library to union the coverage polygons
on the fly and stop if the coverage is close to 100 % for the tile. The
"unified coverage" is then tessellated to find the area.

On many tiles this change will reduce the number of generated and
rendered chart datas by one.