feat(scene): Restructure nodes to support stacked charts
Displaying a seamless chart from multiple chart files requires
simultaneous rendering of those charts in the correct stacking order.

Up until now the following structure has been used to accomplish this:

RootNode
 ┠─PolygonNodeParent
 ┃  ┠─ Tile 1 (polygons for all charts)
 ┃  ┠─ Tile . (polygons for all charts)
 ┃  ┖─ Tile n (polygons for all charts)
 ┃
 ┠─LineNodeParent (not introduced yet)
 ┃    ┠─ Tile 1 (lines for all charts)
 ┃    ┠─ Tile . (lines for all charts)
 ┃    ┖─ Tile n (lines for all charts)
 ┠─SymbolNodeParent
 ┃  ┖─ ...
 ┖─TextNodeParent
    ┖─ ...

The z-values have been assigned so that that polygons are first ordered
based polygon type (land area, built-up area etc.) and then based on the
chart stack.

The problem with this approach is when introducing other rendering
primitives such as lines (e.g polygon borders). A line primitive might
belong to charts down in the stack that should be occluded by polygons
higher up in the stack.

With this structure that is not possible because lines are rendered on
top of all polygons.

The new structure defines GeometryNode that contains the primitives for
each tile such as polygons and lines:

RootNode
 ┠─GeometryNodes
 ┃   ┠─GeometryNode
 ┃   ┃  ┖─┰─ PolygonNode
 ┃   ┃    ┖─ LineNode
 ┃   ┠─GeometryNode .
 ┃   ┃  ┖─┰─ PolygonNode
 ┃   ┃    ┖─ LineNode
 ┃   ┖─GeometryNode n
 ┃     ┖─┰─ PolygonNode
 ┃       ┖─ LineNode
 ┠─SymbolNodeParent
 ┃  ┖─ ...
 ┖─TextNodeParent
    ┖─ ...

Symbol and text rendering is till performed across all charts for each
tile and on top of all geometric primitives.