feat(filefactory): Downsample S-57 data before clipping
Downsampling polygon geometries has up until now been done as part of
the polygon clipping where the values has to be transformed from
floating point positions to integer values anyway. The precision has
been chosen to match the pixel resolution at that zoom level.

Line clipping is done using another algorithm because Clipper2 does
not support line clipping where lines may be split.

There are two problems with this approach which is solved with this
commit:

- Downsampling by mapping to a grid representing the screen pixels
  results in non-optimal data compared to proper line decimation
  algorithms like Ramer–Douglas–Peucker.

- Using different downsampling algorithms for polygon and lines will
  yield different points for the same geographic feature. For example
  polygon area and polygon edge. This can be graphically visible.

In S-57 all polygons and lines are composed of a list of vector edges.
S-57 data is somewhat similar to TopoJSON where polygons and lines
sharing an edge refer to the same vector edge data.

This commit adds downsampling of data at the S-57 level by plugging
into a new vector edge callback in the oesenc library. This callback
is called for every vector edge.

This will ensure that when downsampling at vector edge level the
geometric shapes using those vector edges will be consistent in that for
example corners meet at exactly the same position.

For the actual downsampling a simplify function from boost C++ library
is used in the callback.

For every zoom level a new capnp file is created that includes data for
the entire chart. This "total" capnp file is then used for clipping into
the normal tile files.