feat(tilefactory): Handle both area and line geometry in S57
Many S-57 object types can have both area and line geometries. This
commit refactors the code to handle this in a better way.

The previous line clipping algorithm is replaced with a Cohen-Sutherlands
algorithm from the cutlines library.

Also started to move some of the static methods to anonymous namespace.