build: Define version from git tags
Call `git describe` via CMake to a get version identifier and assign to
the CMake project and expose it though a version.txt file. This will
later be use by the packaging scripts in GHA.

Also added the semantic version to the AppxManifest.xml file. Semantic
versioning doesn't apply directly to the four digit scheme used in MSIX
so it might need more investigation.