ci: Create artifacts using CMake install instead of CPack
CPack is not the solution going forward because deployment to target
systems are seldom done via standalone packages or self-extracting
installers these days.

Another reason to not use CPack is that when downloading artifacts
from GHA they are already being packed as zip so CPack provides a
redundant zip file.

A better approach is to install to temporary directory on the build
agent and let GHA handle the zipping and upload from that directory.

To get the Qt libraries bundled the deploy-qt-windows.cmake script is
re-purposed to a standalone CMake script instead of beeing run through
CPack.

The version string for both Windows and Ubuntu builds is now fetched
from the version.txt file generated in the build folder.