feat(tilefactory): Turn off inflate at chart edges
The "inflate at chart edges" functionality (previously called "move
out edges") was an effort to hide polygon lines along the chart edge.

Those joining lines were a result of stroking amongst others the land
polygons which naturally follow the chart edges unless being completely
inside the chart extent. The inflate functionality would put the line
border outside of the clipping boundary (in the legacy renderer) and
therefore hide it.

I discovered late that the S-57 data format provides specific line object
types such as coast lines and depth contours (not the same as a depth
area) to the needed lines.

As such the inflate functionality is flawed and dead-end attempt to
solve a non-existent problem and it is therefore turned off. The code
can also be removed at later stage.