fix: Fix crash on exit when oexserverd pipe is not connected
If oexserverd is started, but no pipe connection has been established
then the oexserverd process will be killed when nautograf exists and
QProcess::error() will be emitted.

We must not act on this error signal because the CryptReader is in the
the process of being destructed. Solved by disconnecting the
lambdas/slots.