ci: Add Snapcraft core22 build
Eventually the goal is to publish a snap package in the snap store.
But before that more work has to be done to handle the oexserverd
decryption mechanism. Application running as snaps can't just open
unix sockets without additional permissions defined by the package's
metadata.

In the meantime this change will ensure compatibility with the snap
core22 base while continuing development.