feat: Replace decryption management with Catalog class
With the introduction of the updated oesenc library and the Catalog class
a lot of intricacies reading charts can now be handled within the Catalog
class.

This lets us remove the code that was responsible for communication with
oexserverd (cryptreader.cpp). This also means that encrypted chart reading
also should work on Linux now because oesenc library supports it.