refactor(tilefactory): Introduce MeractorTile library
This commit adds a fork of shuair/MercatorTile that can handle much
of the web mercator tile logic code instead of the current homegrown
implementation.

I forecast that the most important tile calculations can be outsourced
to MercatorTile. This will make the code more similar to a standard
web mercator tile renderer and therefore easier to understand.