refactor: Parametrize viewport using center instead of topLeft
This makes many computations easier to understand and it is a more
popular parametrization (used by several web mapping services).
It also has the benefit that when resizing the window it's the center
location which is unchanged and not the top left.

The next step might be to replace pixelsPerLon with a conventional
Web Mercator zoom parameter.