feat: Add xdg-file-dialog for selecting chart catalog
This adds my library xdg-file-dialog to avoid use of
Qt.labs.platform.FolderDialog on Ubuntu 22.04. The build configuration
is set so that the snap package uses xdg-file-dialog, but conventional
build use Qt.labs.platform.FolderDialog.

This change was added to avoid a gtk runtime error which is described
in more detail in the xdg-file-dialog repository.