refactor(scene): Make AnnotationLabel work without symbol
This commit removes "labels" member from AnnotationSymbol. Annotater
class now creates symbols and labels in different arrays.

The QLists also have been replaced with std::vector.

AnotationLabels refer to their parent symbol using an array indexes
instead. Because of this it's important that the original vector does
not change after creation. Sorting is now done in a separate vector with
pointers to the original vector to avoid modification of the original
vector.

This change might improve performance because it removes the plentiful
allocation of 1-element std::vector for almost every symbol. Additionally,
the free standing labels that doesn't need a symbol (for example depth
sounding) are no longer forced to have an empty parent symbol.