feat(scene): Add support for multiple line widths
To achieve this, a uniform is added to the vertex shared for the lines.
Each distinct line width necessitates a unique instance of LineMaterial.
To prevent an excessive number of LineMaterials, the various line widths
are categorized using a new Style struct that defines them as thin,
medium, or thick.

As rendering capabilities expand, additional style properties like dashed
style can be incorporated into this structure

The current lines have been categorized as either thin, medium, or thick
based on their perceived appropriateness.