fix(tilefactory): Avoid holes touching main contour
Investigation showed that the triangulation algorithm failed
to triangulate holes in some rare cases. This was probably
because the hole was touching or intersecting with the
main contour of the polygon.

This commit deflates the all polygon holes with one unit (the
integer coordinate system when clipping). This will reduce or
eliminate the problem.